Refactor Product Cards
