--- src/components/Checkout/MentorshipCard.js
@@ -1,72 +1,0 @@
-import React, { useEffect, useState } from 'react'
-import { useHistory } from 'react-router-dom'
-import { getProfile } from 'services/jwt/index'
-import { Avatar, Rate } from 'antd'
-
-const MentorshipCard = data => {
-  const { listing } = data
-  const [sensei, setSensei] = useState([])
-  const history = useHistory()
-
-  useEffect(() => {
-    getProfile(listing.accountId).then(res => {
-      if (res) {
-        setSensei(res)
-      }
-    })
-  }, [listing.accountId])
-
-  const redirect = id => {
-    history.push({
-      pathname: `/student/mentorship/view/${id}`,
-    })
-  }
-
-  const GetDefaultProfilePic = () => {
-    return '/resources/images/avatars/master.png'
-  }
-
-  return (
-    <div className="col-12">
-      <div
-        role="button"
-        tabIndex={0}
-        className="card btn text-left w-100 "
-        onClick={() => redirect(listing.mentorshipListingId)}
-        onKeyDown={event => event.preventDefault()}
-      >
-        <div className="row">
-          <div className="col-auto">
-            <Avatar
-              size={150}
-              src={
-                sensei.profileImgUrl
-                  ? `${sensei.profileImgUrl}?${new Date().getTime()}`
-                  : GetDefaultProfilePic()
-              }
-            />
-          </div>
-          <div className="col">
-            <h4 className="card-title text-wrap">
-              {sensei.firstName} {sensei.lastName}
-            </h4>
-            <h6 className="card-subtitle mb-2 text-dark text-uppercase text-wrap">
-              {listing.name}
-            </h6>
-            <div className="card-subtitle m-0 text-dark text-wrap">
-              Description : {listing.description}
-            </div>
-            <div className="card-subtitle m-0 text-dark text-wrap">
-              <span className="mt2">
-                <strong>$ {parseFloat(listing.priceAmount).toFixed(2)}</strong>
-              </span>
-            </div>
-            <Rate disabled allowHalf defaultValue={listing.rating} />
-          </div>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-export default MentorshipCard
